added message checks for resolver;refactoring

+++ Broker/handleMessages.go
@@ -135,19 +135,19 @@ func (broker *Broker) validateMessage(message *Message.Message) error {
 	if message.GetSyncResponseToken() != "" {
 		return nil
 	}
-	if broker.config.MaxTopicSize > 0 && uint64(len(message.GetTopic())) > broker.config.MaxTopicSize {
+	if broker.config.MaxTopicSize > 0 && len(message.GetTopic()) > broker.config.MaxTopicSize {
 		return Error.New("message topic is too long", nil)
 	}
-	if broker.config.MaxOriginSize > 0 && uint64(len(message.GetOrigin())) > broker.config.MaxOriginSize {
+	if broker.config.MaxOriginSize > 0 && (len(message.GetOrigin())) > broker.config.MaxOriginSize {
 		return Error.New("message origin is too long", nil)
 	}
-	if broker.config.MaxSyncKeySize > 0 && uint64(len(message.GetSyncRequestToken())) > broker.config.MaxSyncKeySize {
+	if broker.config.MaxSyncKeySize > 0 && (len(message.GetSyncRequestToken())) > broker.config.MaxSyncKeySize {
 		return Error.New("message sync request token is too long", nil)
 	}
-	if broker.config.MaxSyncKeySize > 0 && uint64(len(message.GetSyncResponseToken())) > broker.config.MaxSyncKeySize {
+	if broker.config.MaxSyncKeySize > 0 && (len(message.GetSyncResponseToken())) > broker.config.MaxSyncKeySize {
 		return Error.New("message sync response token is too long", nil)
 	}
-	if broker.config.MaxPayloadSize > 0 && uint64(len(message.GetPayload())) > broker.config.MaxPayloadSize {
+	if broker.config.MaxPayloadSize > 0 && (len(message.GetPayload())) > broker.config.MaxPayloadSize {
 		return Error.New("message payload is too long", nil)
 	}
